refactor : 의존성을 가지는 컴포넌트 같은 디렉토리로 위치 변경
