+++ src/components/domain/BookCardView/BookListSlider/index.js
