+++ src/components/domain/BookCardView/BookListSlider/style.scss
